Implement transform decompiling and use it in patch filter tests

+++ src/wix/WixToolset.Core.WindowsInstaller/WindowsInstallerDecompiler.cs
@@ -3,13 +3,15 @@
 namespace WixToolset.Core.WindowsInstaller
 {
     using System;
-    using System.ComponentModel.Design;
+    using System.Collections.Generic;
     using System.IO;
     using System.Linq;
+    using WixToolset.Core.WindowsInstaller.Bind;
     using WixToolset.Core.WindowsInstaller.Decompile;
     using WixToolset.Core.WindowsInstaller.Unbind;
     using WixToolset.Data;
     using WixToolset.Data.WindowsInstaller;
+    using WixToolset.Extensibility;
     using WixToolset.Extensibility.Data;
     using WixToolset.Extensibility.Services;
 
@@ -22,11 +24,14 @@ namespace WixToolset.Core.WindowsInstaller
         {
             this.ServiceProvider = serviceProvider;
             this.Messaging = serviceProvider.GetService<IMessaging>();
+            this.ExtensionManager = serviceProvider.GetService<IExtensionManager>();
         }
 
         private IServiceProvider ServiceProvider { get; }
 
         private IMessaging Messaging { get; }
+
+        private IExtensionManager ExtensionManager { get; }
 
         public IWindowsInstallerDecompileResult Decompile(IWindowsInstallerDecompileContext context)
         {
@@ -62,7 +67,7 @@ namespace WixToolset.Core.WindowsInstaller
         private IWindowsInstallerDecompileResult Execute(IWindowsInstallerDecompileContext context)
         {
             // Delete the directory and its files to prevent cab extraction failure due to an existing file.
-            if (Directory.Exists(context.ExtractFolder))
+            if (!String.IsNullOrEmpty(context.ExtractFolder) && Directory.Exists(context.ExtractFolder))
             {
                 Directory.Delete(context.ExtractFolder, true);
             }
@@ -71,11 +76,23 @@ namespace WixToolset.Core.WindowsInstaller
 
             var pathResolver = context.ServiceProvider.GetService<IPathResolver>();
 
+            if (context.DecompileType == OutputType.Transform)
+            {
+                return this.DecompileTransform(context, backendHelper, pathResolver);
+            }
+            else
+            {
+                return this.DecompileDatabase(context, backendHelper, pathResolver);
+            }
+        }
+
+        private IWindowsInstallerDecompileResult DecompileDatabase(IWindowsInstallerDecompileContext context, IWindowsInstallerBackendHelper backendHelper, IPathResolver pathResolver)
+        {
             var extractFilesFolder = context.SuppressExtractCabinets || (String.IsNullOrEmpty(context.CabinetExtractFolder) && String.IsNullOrEmpty(context.ExtractFolder)) ? null :
                 String.IsNullOrEmpty(context.CabinetExtractFolder) ? Path.Combine(context.ExtractFolder, "File") : context.CabinetExtractFolder;
 
             var outputType = context.TreatProductAsModule ? OutputType.Module : context.DecompileType;
-            var unbindCommand = new UnbindDatabaseCommand(this.Messaging, backendHelper, pathResolver, context.DecompilePath, outputType, context.ExtractFolder, extractFilesFolder, context.IntermediateFolder, enableDemodularization: true, skipSummaryInfo: false);
+            var unbindCommand = new UnbindDatabaseCommand(this.Messaging, backendHelper, pathResolver, context.DecompilePath, null, outputType, context.ExtractFolder, extractFilesFolder, context.IntermediateFolder, enableDemodularization: true, skipSummaryInfo: false);
             var output = unbindCommand.Execute();
             var extractedFilePaths = unbindCommand.ExportedFiles;
 
@@ -84,10 +101,24 @@ namespace WixToolset.Core.WindowsInstaller
             var document = decompiler.Decompile(output);
 
             var result = context.ServiceProvider.GetService<IWindowsInstallerDecompileResult>();
+            result.Data = output;
             result.Document = document;
             result.Platform = GetPlatformFromOutput(output);
             result.ExtractedFilePaths = extractedFilePaths.ToList();
+            return result;
+        }
 
+        private IWindowsInstallerDecompileResult DecompileTransform(IWindowsInstallerDecompileContext context, IWindowsInstallerBackendHelper backendHelper, IPathResolver pathResolver)
+        {
+            var fileSystemExtensions = this.ExtensionManager.GetServices<IFileSystemExtension>();
+
+            var fileSystemManager = new FileSystemManager(fileSystemExtensions);
+
+            var unbindCommand = new UnbindTransformCommand(this.Messaging, backendHelper, pathResolver, fileSystemManager, context.DecompilePath, context.ExtractFolder, context.IntermediateFolder);
+            var output = unbindCommand.Execute();
+
+            var result = context.ServiceProvider.GetService<IWindowsInstallerDecompileResult>();
+            result.Data = output;
             return result;
         }
